Give items of non-persistent containers back to the player upon close.
Before this commit, all items in the inventory of non-persistent
custom containers would be dropped on the ground when the player
closes it (or is forced to close it for some reason).

This behavior is changed by this commit: the plug-in will try to
fit as many items of the custom container in the inventory of the
player, and only drop the items that don't fit. This is more
in-line with the behavior of crafting tables.