Revert "Search for corrupted item stacks"
This reverts commit 9295229a01c1697b9ababc11162a7c98e05c32e8.

The 'corrupted item stack bug' should be fixed in the next
build of Spigot/Paper.

I am using quotation marks because it turned out that the
item stacks were not corrupted at all: the Spigot
implementation of AttributeModifier.getUniqueId() would
simply throw this exception for every ItemStack.

The exception was only thrown after updating the item set
on a server, because only then would this plug-in attempt
to read the attribute modifier( id)s.